requirements.txt
Flask
joblib
pandas
numpy
scikit-learn==1.3.2
xgboost==1.7.6
tldextract
requests
gunicorn
python-whois
